requirements.txt
Django==5.0.6
djangorestframework==3.15.1
PyMySQL==1.1.1
redis==5.0.6
requests==2.32.3
gunicorn==22.0.0
django-redis==5.4.0
